choix de la page pour un projet client

+++ client/src/admin/partners/list.js
@@ -35,40 +35,44 @@ const ListElement = ({ slug, fields }) => {
 
             <SectionTitle title={ t(`${slugTrans}.label`) } subtitle={ t(`${slugTrans}.all`) } />
 
-            { items ? items.map((type, i) => (
-                <Widget key={i} title={type.label} description={`Tous les partenaires du groupe : ${type.label}`}>
-                    <Link to={ {pathname: `/admin/${slug}/new`, state: {partner_type: type.id} }} className="btn btn-sm bg-blue-500 hover:bg-blue-600 text-white btn-rounded mb-2">
-                        <span>{ t(`${slugTrans}.new`) }</span>
-                    </Link>
-                    <table className="table no-border striped">
-                        <thead>
-                            <tr> 
-                                { fields.map((field, i) => (
-                                    <th key={ i }> { field.name } </th>
-                                ))}
-                            </tr> 
-                        </thead>
-                        <tbody>
-                            { type.partners ? type.partners.map((item, i) => ( 
-                                <tr key={ i }> 
-                                    { fields.map((field, j) => (
-                                        <td key={ j }> { item[field.key] } </td>
-                                    ))} 
-                                    <td className="text-right">
-                                        <Link className="btn btn-circle bg-transparent hover:bg-blue-50 text-blue-500 hover:text-blue-600 btn-raised"
-                                        title="Modifier l'article" to={ { pathname: `/admin/${slug}/update/${item['id']}`, state: { partner_type: type.id } } }>
-                                        <FiEdit className = "stroke-current" /></Link>
-                                        <button className="btn btn-circle bg-transparent hover:bg-blue-50 text-blue-500 hover:text-blue-600 btn-raised ml-3"
-                                        title="Supprimer l'article" onClick={() => { deleteItem(item['id']) }}>
-                                        <FiDelete className="stroke-current" />
-                                        </button>
-                                    </td>
-                                </tr>
-                            )) : null }
-                        </tbody>
-                    </table>
-                </Widget> 
-            )) : ''}
+            { items ? items.map((type, i) => {  
+                if (type.label !== 'Projets clients') {
+                    return (
+                        <Widget key={i} title={type.label} description={`Tous les partenaires du groupe : ${type.label}`}>
+                            <Link to={ {pathname: `/admin/${slug}/new`, state: {partner_type: type.id} }} className="btn btn-sm bg-blue-500 hover:bg-blue-600 text-white btn-rounded mb-2">
+                                <span>{ t(`${slugTrans}.new`) }</span>
+                            </Link>
+                            <table className="table no-border striped">
+                                <thead>
+                                    <tr> 
+                                        { fields.map((field, i) => (
+                                            <th key={ i }> { field.name } </th>
+                                        ))}
+                                    </tr> 
+                                </thead>
+                                <tbody>
+                                    { type.partners ? type.partners.map((item, i) => ( 
+                                        <tr key={ i }> 
+                                            { fields.map((field, j) => (
+                                                <td key={ j }> { item[field.key] } </td>
+                                            ))} 
+                                            <td className="text-right">
+                                                <Link className="btn btn-circle bg-transparent hover:bg-blue-50 text-blue-500 hover:text-blue-600 btn-raised"
+                                                title="Modifier l'article" to={ { pathname: `/admin/${slug}/update/${item['id']}`, state: { partner_type: type.id } } }>
+                                                <FiEdit className = "stroke-current" /></Link>
+                                                <button className="btn btn-circle bg-transparent hover:bg-blue-50 text-blue-500 hover:text-blue-600 btn-raised ml-3"
+                                                title="Supprimer l'article" onClick={() => { deleteItem(item['id']) }}>
+                                                <FiDelete className="stroke-current" />
+                                                </button>
+                                            </td>
+                                        </tr>
+                                    )) : null }
+                                </tbody>
+                            </table>
+                        </Widget> 
+                    )
+                }
+            }) : ''}
         </>
     )
 }
